handle DB-transfers better

+++ daemon97.py
@@ -104,6 +104,8 @@ def do_writesample(cnsql, cmd, sample):
       print "***** Closing cursor"
       cursql.close()
     print(e.__doc__)
+    syslog.syslog(syslog.LOG_ALERT,e.__doc__)
+    syslog_trace(traceback.format_exc())
 
 def do_sql_data(flock, inicnfg, cnsql):
   if DEBUG:print "Pushing data to MySQL-server"
@@ -120,34 +122,44 @@ def do_sql_data(flock, inicnfg, cnsql):
   #endwhile
 
   for inisect in inicnfg.sections(): # Check each section of the config.ini file
+    errsql = False
     try:
       ifile = inicnfg.get(inisect,"resultfile")
       if DEBUG:print ifile
+
+      try:
+        sqlcmd = []
+        sqlcmd = inicnfg.get(inisect,"sqlcmd")
+
+        try:
+          data = cat(ifile).splitlines()
+          if (len(data) > 0):
+            for entry in range(0, len(data)):
+              if DEBUG:print data[entry]
+              do_writesample(cnsql, sqlcmd, data[entry])
+            #endfor
+          #endif
+        except:
+          errsql = True
+          if DEBUG:print "Error while attempting to write data to DB"
+          syslog.syslog(syslog.LOG_ALERT,e.__doc__)
+          syslog_trace(traceback.format_exc())
+
+      except:
+        if DEBUG:print "No SQL command defined for section", inisect
+
     except:
       if DEBUG:print "No resultfile for section", inisect
-    sqlcmd = []
-    try:
-      sqlcmd = inicnfg.get(inisect,"sqlcmd")
-    except:
-      if DEBUG:print "No SQL command defined for section", inisect
 
-    if (sqlcmd != []):
-      data = cat(ifile).splitlines()
-      if (len(data) > 0):
-        for entry in range(0, len(data)):
-          if DEBUG:print data[entry]
-          do_writesample(cnsql, sqlcmd, data[entry])
-        #endfor
-      #endif
-    #endif
     try:
       ofile = inicnfg.get(inisect,"rawfile")
       if DEBUG:print ofile
-      if os.path.isfile(ifile):       # resultfile exists
-        if not os.path.isfile(ofile): # rawfile does not exist
-          shutil.move(ifile, ofile)   # then move the file over
+      if not errsql:                    # SQL-job was successful or non-existing
+        if os.path.isfile(ifile):       # resultfile exists
+          if not os.path.isfile(ofile): # rawfile does not exist
+            shutil.move(ifile, ofile)   # then move the file over
     except:
-      if DEBUG:print "No rawfile defined for section", inisect
+      if DEBUG:print "No rawfile defined for section or error while moving", inisect
   #endfor
   unlock(flock)
